+++ src/openassetio-core/tests/trait/TraitBaseTest.cpp
@@ -56,12 +56,31 @@ SCENARIO("Retrieving the undelying data") {
 }
 
 SCENARIO("Checking a trait is imbued") {
-  GIVEN("a trait view on an empty TraitsData") {
+  GIVEN("an empty TraitsData") {
     const TraitsDataPtr data = TraitsData::make();
-    const TestTrait trait(data);
 
-    WHEN("the view is queried for the imbued status") {
-      const bool isImbued = trait.isImbued();
+    AND_GIVEN("an instance of a trait view wrapping the TraitsData") {
+      const TestTrait trait(data);
+
+      WHEN("the view instance is queried for the imbued status") {
+        const bool isImbued = trait.isImbued();
+
+        THEN("the view reports that the trait is not imbued") { CHECK(!isImbued); }
+      }
+
+      AND_GIVEN("the TraitsData is imbued with the view's trait") {
+        data->addTrait(TestTrait::kId);
+
+        WHEN("the view instance is queried for the imbued status") {
+          const bool isImbued = trait.isImbued();
+
+          THEN("the view reports that the trait is imbued") { CHECK(isImbued); }
+        }
+      }
+    }
+
+    WHEN("the TraitsData is queried for imbued status using static trait view function") {
+      const bool isImbued = TestTrait::isImbuedTo(data);
 
       THEN("the view reports that the trait is not imbued") { CHECK(!isImbued); }
     }
@@ -69,8 +88,8 @@ SCENARIO("Checking a trait is imbued") {
     AND_GIVEN("the TraitsData is imbued with the view's trait") {
       data->addTrait(TestTrait::kId);
 
-      WHEN("the view is queried for the imbued status") {
-        const bool isImbued = trait.isImbued();
+      WHEN("the TraitsData is queried for imbued status using static trait view function") {
+        const bool isImbued = TestTrait::isImbuedTo(data);
 
         THEN("the view reports that the trait is imbued") { CHECK(isImbued); }
       }