+++ src/openassetio-python/tests/package/test_trait.py
@@ -36,11 +36,13 @@ class Test_Trait_Construction:
 class Test_Trait_isImbued:
     def test_when_data_has_trait_returns_true(self):
         a_data = TraitsData({ACustomTrait.kId})
+        assert ACustomTrait.isImbuedTo(a_data) is True
         trait = ACustomTrait(a_data)
         assert trait.isImbued() is True
 
     def test_when_data_does_not_have_trait_returns_false(self):
         a_data = TraitsData({"someOtherTrait"})
+        assert ACustomTrait.isImbuedTo(a_data) is False
         trait = ACustomTrait(a_data)
         assert trait.isImbued() is False