+++ RELEASE_NOTES.md
@@ -10,6 +10,12 @@ v1.0.0-alpha.X
   `imbue`/`imbueTo` methods.
   [#815](https://github.com/OpenAssetIO/OpenAssetIO/issues/815)
 
+### New features
+
+- Added `TraitBase.isImbuedTo` static/class method, giving a cheaper
+  mechanism for testing whether a `TraitsData` is imbued with a trait.
+  [#815](https://github.com/OpenAssetIO/OpenAssetIO/issues/815)
+
 
 v1.0.0-alpha.9
 --------------